Refactoring: Fixed some warnings.

+++ src/Bitmap.cpp
@@ -39,7 +39,7 @@ void Bitmap::update() {
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_S, GL_REPEAT);
     glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_WRAP_T, GL_REPEAT);
 
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, m_width, m_height, 0, GL_RGBA, GL_UNSIGNED_BYTE, m_pixels.get());
+    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA, static_cast<int>(m_width), static_cast<int>(m_height), 0, GL_RGBA, GL_UNSIGNED_BYTE, m_pixels.get());
 
     m_has_changed = false;
 }
